+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Changed
 
 - the semaphore used to throttle the accept loop is now also used for outbound connections
+- the MSRV is bumped to 1.88
 
 ### Fixed